// Rainmaker2/src/Presentation/colaba-admin/Settings/src/Utils/helpers/Enums.ts
export enum ToolTipEnum {
    DocumentSubmit = 1,
    LoanApplicationSubmitted = 2,
    LoanFunded = 3,
    NotificationHeader = 4,
    ManageDocumentTemplateHeader = 5,
    SortOrder = 6,
    TokenList = 7,
    ByteSoftInteg = 8,
    RequestEmailTemplates = 9,
    LoanStatusUpdate = 10,
    SendEmailAfter = 11
}

export enum ProfileMenu {
    AccountDetail = 1,
    AssignedRole = 2,
    Notification = 3
}

export enum LoadOriginationMenu {
    Users = 1,
    Organization = 2
}
export enum SimpleSort{
    Up =1,
    Down=2
}